add model variations table and ToC
Following discussion at https://huggingface.co/bert-base-uncased/discussions/6 I added a "Model variations" section to the model card, it has a brief history of variations with a link to the BERT github readme for detailed info. A table reports the relevant models on the HF hub.
I also added a ToC on the top, as seen in camemBERT for example.