Added model variations section
Following https://huggingface.co/bert-base-uncased/discussions/6 I added a "model variations" section to the model card, it has a brief history of variations with a link to the BERT github readme for detailed info. A table reports the relevant models on the HF hub. 
Copy of closed PR https://huggingface.co/bert-base-uncased/discussions/8 but without the added ToC as suggested by @julien-c